HadithHouseWebsite/requirements.txt
Django==1.9.3
django-crispy-forms==1.6.0
django-filter==0.12.0
djangorestframework==3.3.2
mock==2.0.0
psycopg2==2.6.1
sqlparse==0.2.2
urlfetch==1.0.2